Disable confirm multiline paste dialog
On newer versions of QTermWidget (0.17) when trying to paste multiline
text annoying querry dialog appears. In this commit this is disabled, so
there will be no query dialog.

However, since this function is not available on older version of
qtermwidget (which does not query anything by default) it is added here
as commented out line, which can be uncommented when using qtermwidget
0.17 or higher.