Add ability to increase/decrease font
1 - Add ability to increase or decrease font
2 - Keep focus on terminal after "sync with editor" click
3 - Change shortcuts for copy and paste to
Ctrl+Shift+C and Ctrl+Shift+V, which matches linux
terminal shortcuts, since Ctrl+C is used to cancel
current execution in terminal.
Note that QtCreator already has this shortcuts taken for
come other actions, and they need to be reset.
Will see in future if it is possible to change this in smarter way.